requirements.txt
aniso8601==7.0.0
astroid==2.6.2
attrs==21.2.0
autopep8==1.5.7
Django==2.2.13
django-cors-headers==3.7.0
django-filter==2.3.0
django-model-utils==4.0.0
flake8==3.9.2
flake8-isort==4.0.0
graphene==2.1.8
graphene-django==2.10.1
graphql-core==2.3.2
graphql-relay==2.0.1
iniconfig==1.1.1
isort==5.9.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
packaging==21.0
pluggy==0.13.1
promise==2.3
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.9.3
pylint-celery==0.3
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.2.4
pytest-sugar==0.9.4
python-dotenv==0.13.0
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
termcolor==1.1.0
testfixtures==6.17.1
toml==0.10.2
wrapt==1.12.1
